// lib/database.types.ts
export interface Database {
  public: {
    Tables: {
      profiles: {
        Row: {
          id: string
          username: string
          full_name: string | null
          avatar_url: string | null
          bio: string | null
          account_type: string
          subscription_price: number | null
          created_at: string
          updated_at: string
          email: string
        }
        Insert: {
          id: string
          username: string
          full_name?: string | null
          avatar_url?: string | null
          bio?: string | null
          account_type: string
          subscription_price?: number | null
          created_at?: string
          updated_at?: string
          email: string
        }
        Update: {
          id?: string
          username?: string
          full_name?: string | null
          avatar_url?: string | null
          bio?: string | null
          account_type?: string
          subscription_price?: number | null
          created_at?: string
          updated_at?: string
          email?: string
        }
      }
      posts: {
        Row: {
          id: string
          title: string
          content: string | null
          file_url: string | null
          creator_id: string
          is_preview: boolean
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: string
          title: string
          content?: string | null
          file_url?: string | null
          creator_id: string
          is_preview?: boolean
          created_at?: string
          updated_at?: string
        }
        Update: {
          id?: string
          title?: string
          content?: string | null
          file_url?: string | null
          creator_id?: string
          is_preview?: boolean
          created_at?: string
          updated_at?: string
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}